requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
pillow-avif-plugin==1.4.2
pillow-heif
numpy==1.24.3
mediapipe==0.10.8
pytest>=7.0.0
